refactor: fix wrong dir naming for graphql federation samples
