+++ sample/31-graphql-federation-code-first/posts-application/src/posts/models/post.model.ts
@@ -1,0 +1,18 @@
+import { Directive, Field, ID, Int, ObjectType } from '@nestjs/graphql';
+import { User } from './user.model';
+
+@ObjectType()
+@Directive('@key(fields: "id")')
+export class Post {
+  @Field((type) => ID)
+  id: number;
+
+  @Field()
+  title: string;
+
+  @Field((type) => Int)
+  authorId: number;
+
+  @Field((type) => User)
+  user?: User;
+}
