+++ sample/31-graphql-federation-code-first/posts-application/src/posts/models/user.model.ts
@@ -1,0 +1,14 @@
+import { Directive, Field, ID, ObjectType } from '@nestjs/graphql';
+import { Post } from './post.model';
+
+@ObjectType()
+@Directive('@extends')
+@Directive('@key(fields: "id")')
+export class User {
+  @Field((type) => ID)
+  @Directive('@external')
+  id: number;
+
+  @Field((type) => [Post])
+  posts?: Post[];
+}
