+++ sample/31-graphql-federation-code-first/posts-application/src/posts/posts.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { GraphQLFederationModule } from '@nestjs/graphql';
+import { User } from './models/user.model';
 import { PostsResolver } from './posts.resolver';
 import { PostsService } from './posts.service';
 import { UsersResolver } from './users.resolver';
@@ -7,7 +8,10 @@ import { UsersResolver } from './users.resolver';
 @Module({
   imports: [
     GraphQLFederationModule.forRoot({
-      typePaths: ['**/*.graphql'],
+      autoSchemaFile: true,
+      buildSchemaOptions: {
+        orphanedTypes: [User],
+      },
     }),
   ],
   providers: [PostsService, PostsResolver, UsersResolver],
