+++ sample/31-graphql-federation-code-first/posts-application/src/posts/posts.resolver.ts
@@ -6,24 +6,25 @@ import {
   Parent,
   ID,
 } from '@nestjs/graphql';
-import { Post } from './posts.interfaces';
+import { Post } from './models/post.model';
+import { User } from './models/user.model';
 import { PostsService } from './posts.service';
 
-@Resolver()
+@Resolver((of) => Post)
 export class PostsResolver {
   constructor(private readonly postsService: PostsService) {}
 
-  @Query()
-  findPost(@Args({ name: 'id', type: () => ID }) id: number) {
+  @Query((returns) => Post)
+  findPost(@Args({ name: 'id', type: () => ID }) id: number): Post {
     return this.postsService.findOne(id);
   }
 
-  @Query()
-  getPosts() {
+  @Query((returns) => [Post])
+  getPosts(): Post[] {
     return this.postsService.all();
   }
 
-  @ResolveField('user')
+  @ResolveField((of) => User)
   user(@Parent() post: Post): any {
     return { __typename: 'User', id: post.authorId };
   }
