+++ sample/31-graphql-federation-code-first/posts-application/src/posts/posts.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { Post } from './posts.interfaces';
+import { Post } from './models/post.model';
 
 @Injectable()
 export class PostsService {
