+++ sample/31-graphql-federation-code-first/users-application/src/users/models/user.model.ts
@@ -1,0 +1,11 @@
+import { Directive, Field, ID, ObjectType } from '@nestjs/graphql';
+
+@ObjectType()
+@Directive('@key(fields: "id")')
+export class User {
+  @Field((type) => ID)
+  id: number;
+
+  @Field()
+  name: string;
+}
