+++ sample/31-graphql-federation-code-first/users-application/src/users/users.module.ts
@@ -7,7 +7,7 @@ import { UsersService } from './users.service';
   providers: [UsersResolver, UsersService],
   imports: [
     GraphQLFederationModule.forRoot({
-      typePaths: ['**/*.graphql'],
+      autoSchemaFile: true,
     }),
   ],
 })
