+++ sample/31-graphql-federation-code-first/users-application/src/users/users.resolver.ts
@@ -1,17 +1,18 @@
 import { Args, ID, Query, Resolver, ResolveReference } from '@nestjs/graphql';
+import { User } from './models/user.model';
 import { UsersService } from './users.service';
 
-@Resolver('User')
+@Resolver((of) => User)
 export class UsersResolver {
   constructor(private usersService: UsersService) {}
 
-  @Query()
-  getUser(@Args({ name: 'id', type: () => ID }) id: number) {
+  @Query((returns) => User)
+  getUser(@Args({ name: 'id', type: () => ID }) id: number): User {
     return this.usersService.findById(id);
   }
 
   @ResolveReference()
-  resolveReference(reference: { __typename: string; id: number }) {
+  resolveReference(reference: { __typename: string; id: number }): User {
     return this.usersService.findById(reference.id);
   }
 }
