+++ sample/31-graphql-federation-code-first/users-application/src/users/users.service.ts
@@ -1,13 +1,14 @@
 import { Injectable } from '@nestjs/common';
+import { User } from './models/user.model';
 
 @Injectable()
 export class UsersService {
-  private users = [
+  private users: User[] = [
     { id: 1, name: 'John Doe' },
     { id: 2, name: 'Richard Roe' },
   ];
 
-  findById(id: number) {
+  findById(id: number): User {
     return this.users.find((user) => user.id === Number(id));
   }
 }
