+++ sample/32-graphql-federation-schema-first/posts-application/src/posts/posts.graphql
@@ -2,6 +2,7 @@ type Post @key(fields: "id") {
   id: ID!
   title: String!
   body: String!
+  authorId: Int!
   user: User
 }
