+++ sample/32-graphql-federation-schema-first/posts-application/src/posts/posts.interfaces.ts
