+++ sample/32-graphql-federation-schema-first/posts-application/src/posts/posts.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { GraphQLFederationModule } from '@nestjs/graphql';
-import { User } from './models/user.model';
 import { PostsResolver } from './posts.resolver';
 import { PostsService } from './posts.service';
 import { UsersResolver } from './users.resolver';
@@ -8,10 +7,7 @@ import { UsersResolver } from './users.resolver';
 @Module({
   imports: [
     GraphQLFederationModule.forRoot({
-      autoSchemaFile: true,
-      buildSchemaOptions: {
-        orphanedTypes: [User],
-      },
+      typePaths: ['**/*.graphql'],
     }),
   ],
   providers: [PostsService, PostsResolver, UsersResolver],
