+++ sample/32-graphql-federation-schema-first/posts-application/src/posts/users.interfaces.ts
