+++ sample/32-graphql-federation-schema-first/users-application/src/users/users.graphql
