+++ sample/32-graphql-federation-schema-first/users-application/src/users/users.resolver.ts
@@ -1,18 +1,17 @@
 import { Args, ID, Query, Resolver, ResolveReference } from '@nestjs/graphql';
-import { User } from './models/user.model';
 import { UsersService } from './users.service';
 
-@Resolver((of) => User)
+@Resolver('User')
 export class UsersResolver {
   constructor(private usersService: UsersService) {}
 
-  @Query((returns) => User)
-  getUser(@Args({ name: 'id', type: () => ID }) id: number): User {
+  @Query()
+  getUser(@Args({ name: 'id', type: () => ID }) id: number) {
     return this.usersService.findById(id);
   }
 
   @ResolveReference()
-  resolveReference(reference: { __typename: string; id: number }): User {
+  resolveReference(reference: { __typename: string; id: number }) {
     return this.usersService.findById(reference.id);
   }
 }
